int -> size_t

+++ Channel.cpp
@@ -26,7 +26,7 @@ bool	Channel::is_valid_channel(string channel) {
 	if (channel[0] != '&' && channel[0] != '#'
 		&& channel[0] != '+' && channel[0] != '!')
 		return (false);
-	for (unsigned long i = 1; i < channel.size(); i++) {
+	for (size_t i = 1; i < channel.size(); i++) {
 		if (channel[i] == ' ' || channel[i] == 7 || channel[i] == ',')
 			return (false);
 	}
@@ -36,7 +36,7 @@ bool	Channel::is_valid_channel(string channel) {
 bool	Channel::is_valid_pass(string password) {
 	if (password.empty() || password.size() > 23)
 		return (false);
-	for (unsigned long i = 0; i < password.size(); i++) {
+	for (size_t i = 0; i < password.size(); i++) {
 		if (password[i] == '\0' || password[i] == '\r'
 			|| std::isspace(password[i]))
 			return (false);
